knex delete table only if exists

+++ packages/server-graphql/src/SchemaKnex/SchemaKnex_DeleteTable.ts
@@ -4,10 +4,15 @@ export default function DeleteTable(ioc: IIOC) {
   const { knex } = ioc;
 
   return function (model: IModel) {
-    return async function (): Promise<void> {
+    return async function (): Promise<null | false> {
       const { name } = model;
+
+      if (!(await knex.schema.hasTable(name))) {
+        return false;
+      }
+
       await knex.schema.dropTable(name);
-      return;
+      return null;
     };
   };
 }
